[refactor] 拆分 comments controllers 結構
