+++ controllers/comments/index.js
@@ -1,0 +1,5 @@
+const commentBase = require("./commentBase");
+
+module.exports = {
+  ...commentBase,
+};
