+++ routes/user/commentRoutes.js
@@ -1,18 +1,20 @@
 const express = require("express");
 const router = express.Router();
-const CommentControllers = require("../../controllers/user/comments");
+const comment = require("../../controllers/comments");
 const errorAsyncHandler = require("../../services/errorAsyncHandler");
 const {
   checkTokenAndSetAuth,
   getUserFromAuthId,
 } = require("../../middleware/authMiddleware");
 
+// commentBase
+
 // 刪除評論
 router.delete(
   "/comment/:commentId",
   checkTokenAndSetAuth,
   getUserFromAuthId,
-  errorAsyncHandler(CommentControllers.deleteComment)
+  errorAsyncHandler(comment.deleteComment)
 );
 
 module.exports = router;
